Lot of changes
Working with the output from Assembly, trying to run it in Basic. Converted Basic to a rough approximation of VBScript. Appears to be functionally correct yet throws issues with the CHR function. Circled back to Assembly to try and sort out if the output I received from it was correct. Found SASM (Assembly debugger) and need to explore assembly more to understand if the output is correct. Will be cleaning this up in a future commit.